backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
google-generativeai==0.4.0
ddtrace==2.6.0
datadog-api-client==2.22.0
python-dotenv==1.0.0
pydantic==2.5.3
httpx==0.26.0
python-multipart==0.0.6
aiofiles==23.2.1
structlog==24.1.0
